fixed discard bug

+++ app/Http/Controllers/ReportController.php
@@ -10,22 +10,19 @@ class ReportController extends Controller
 {
     public function report_review(Request $request, $id){
 
-        $this->authorize('report', [Review::class, $id]);
+        $this->authorize('report', [Review::class, Review::find($id)]);
 
         $request->user()->reported()->create([
             'review_id' => $id
         ]);
-
-        return back();
     }
 
     public function discard(Request $request, $id){
+        $this->authorize('discardReport',[Report::class,$request->user()]);
 
         $reports = Report::where('review_id', $id);
 
         $reports->delete();   
-
-        return back();
     }
 
 }
